Basic tile fires impl (#483)
* Basic dumb not really working tile fire impl

* truly cannot be assed to do all of these like 90% of them are unused on toast anyway

* tile fire spread logic

* visual stuff again and stacks

* ThatOneGamingArsonist

* tweak and juice merchant

* buff

* poop api

* welcome to spark fire city

* triggerslop

* tweaking

* i guess lol

(cherry picked from commit 4444a6935349c9928db5a6c99454c18713a2733d)

+++ Content.Client/Atmos/EntitySystems/FireVisualizerSystem.cs
@@ -64,7 +64,10 @@ public sealed class FireVisualizerSystem : VisualizerSystem<FireVisualsComponent
             return;
 
         AppearanceSystem.TryGetData<bool>(uid, FireVisuals.OnFire, out var onFire, appearance);
-        AppearanceSystem.TryGetData<float>(uid, FireVisuals.FireStacks, out var fireStacks, appearance);
+        // ES START
+        // firestacks floored to int instead of float
+        AppearanceSystem.TryGetData<int>(uid, FireVisuals.FireStacks, out var fireStacks, appearance);
+        // ES END
         SpriteSystem.LayerSetVisible((uid, sprite), index, onFire);
 
         if (!onFire)
